Move hooks to a folder.
